requirements.txt
flask
fastjsonschema
psycopg2-binary